refactor(python-keymaster): move gatekeeper_client and wallet_store into keymaster library
These utilities are library-level concerns (not service-specific), so
moving them lets keymaster_cli depend on archon-keymaster alone instead
of pulling in the full keymaster_service package. keymaster_service now
imports them from the keymaster library via absolute imports.

- Adds redis dependency to archon-keymaster (required by wallet_store)
- Removes keymaster-service dependency from keymaster-cli
- Updates README install instructions
