+++ python/keymaster/pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
   "bolt11==2.2.0",
   "coincurve==21.0.0",
   "cryptography==46.0.5",
+  "redis==5.2.1",
   "rfc8785==0.1.4"
 ]
