+++ python/keymaster_cli/pyproject.toml
@@ -8,8 +8,7 @@ version = "0.1.0"
 description = "Python keymaster CLI mirroring the TypeScript keymaster CLI"
 readme = "README.md"
 dependencies = [
-  "archon-keymaster==0.1.0",
-  "keymaster-service==0.1.0"
+  "archon-keymaster==0.1.0"
 ]
 
 [project.scripts]
